move tomato field logic to utils
