+++ blitz_api/models.py
@@ -2,7 +2,6 @@ import binascii
 import datetime
 import os
 import logging
-import random
 
 from django.conf import settings
 from django.db import models
@@ -18,6 +17,7 @@ from simple_history.models import HistoricalRecords
 from . import mailchimp
 
 from tomato.models import Tomato
+from utils.tomato_field import TomatoFieldManager
 from blitz_api import services
 from blitz_api.managers import ActionTokenManager
 
@@ -39,90 +39,6 @@ class User(AbstractUser):
         (LANGUAGE_EN, _('English')),
         (LANGUAGE_FR, _('French')),
     )
-
-    TOMATO_MATRIX_NUMBER_OF_ANIMAL = 1
-    TOMATO_MATRIX_NUMBER_OF_DECORATION = 2
-
-    TOMATO_MATRIX_CELLS_ANIMAL = [
-        'grass_pig_2',
-        'grass_cow',
-        'grass_horse',
-        'grass_goat',
-        'grass_dog',
-        'grass_chicken',
-        'grass_pig',
-        'grass_sheep_2',
-        'grass_sheep',
-    ]
-    TOMATO_MATRIX_CELLS_DECORATION = [
-        'grass_well',
-        'grass_straw',
-        'grass_wood_2',
-        'grass_wood',
-        'grass_lawn',
-        'grass_tree_2',
-        'grass_tree',
-    ]
-    TOMATO_MATRIX_CELLS_DEFAULT = 'grass'
-    TOMATO_MATRIX_CELLS_DEFAULT_FIELD = 'field_0'
-    TOMATO_MATRIX_CELLS_BENCH = 'grass_thv'
-
-    TOMATO_MATRIX_CELLS = [
-        *TOMATO_MATRIX_CELLS_DEFAULT,
-        *TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-        *TOMATO_MATRIX_CELLS_BENCH,
-        *TOMATO_MATRIX_CELLS_DECORATION,
-        *TOMATO_MATRIX_CELLS_ANIMAL,
-    ]
-
-    TOMATO_MATRIX_BASE = [
-        [
-            TOMATO_MATRIX_CELLS_DEFAULT,
-            TOMATO_MATRIX_CELLS_DEFAULT,
-            TOMATO_MATRIX_CELLS_DEFAULT,
-            TOMATO_MATRIX_CELLS_DEFAULT,
-            TOMATO_MATRIX_CELLS_DEFAULT,
-        ],
-        [
-            TOMATO_MATRIX_CELLS_DEFAULT,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-        ],
-        [
-            TOMATO_MATRIX_CELLS_DEFAULT,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-        ],
-        [
-            TOMATO_MATRIX_CELLS_DEFAULT,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-        ],
-        [
-            TOMATO_MATRIX_CELLS_BENCH,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-            TOMATO_MATRIX_CELLS_DEFAULT_FIELD,
-        ],
-    ]
-
-    TOMATO_MATRIX_RANDOM_CELLS = [
-        (0, 0),
-        (0, 1),
-        (0, 2),
-        (0, 3),
-        (0, 4),
-        (1, 0),
-        (2, 0),
-        (3, 0),
-    ]
 
     phone = models.CharField(
         verbose_name=_("Phone number"),
@@ -506,53 +422,12 @@ class User(AbstractUser):
         self.tickets += nb_tickets
         self.save()
 
-    def generate_tomato_field_matrix(self):
-        """
-        Generate a matrix with personalised cells
-        return the generated matrix as a list of list
-        """
-        matrix = self.TOMATO_MATRIX_BASE
-
-        # Decide which cell should be personalised
-        total_number_of_personalisation = \
-            self.TOMATO_MATRIX_NUMBER_OF_DECORATION + \
-            self.TOMATO_MATRIX_NUMBER_OF_ANIMAL
-
-        cell_to_personalise = random.sample(
-            self.TOMATO_MATRIX_RANDOM_CELLS,
-            total_number_of_personalisation,
-        )
-
-        cell_count = 0
-
-        # Personalisation of animals
-        random_animals = random.sample(
-            self.TOMATO_MATRIX_CELLS_ANIMAL,
-            self.TOMATO_MATRIX_NUMBER_OF_ANIMAL,
-        )
-        for animal in random_animals:
-            cell = cell_to_personalise[cell_count]
-            matrix[cell[0]][cell[1]] = animal
-            cell_count += 1
-
-        # Personalisation of decorations
-        random_decorations = random.sample(
-            self.TOMATO_MATRIX_CELLS_DECORATION,
-            self.TOMATO_MATRIX_NUMBER_OF_DECORATION,
-        )
-        for decoration in random_decorations:
-            cell = cell_to_personalise[cell_count]
-            matrix[cell[0]][cell[1]] = decoration
-            cell_count += 1
-
-        return matrix
-
     @property
     def tomato_field_matrix(self):
         if self._tomato_field_matrix:
             return self._tomato_field_matrix
         else:
-            matrix = self.generate_tomato_field_matrix()
+            matrix = TomatoFieldManager.generate_tomato_field_matrix()
             self._tomato_field_matrix = matrix
             self.save()
             return matrix
